Add files via upload
In this lab I extract the information of the top 50 movies with the best average rating from the web link shared below.
https://web.archive.org/web/20230902185655/https://en.everybodywiki.com/100_Most_Highly-Ranked_Films

The information required is Average Rank, Film, and Year.
Here I write a Python script that extracts the information and saves it to a CSV file top_50_films.csv. And I also save the same information to a database Movies.db under the table name Top_50.